refactor(member) memeber 패키지 구조 변경
