+++ member-service/src/main/java/com/green/member/application/admin/AdminController.java
@@ -1,6 +1,6 @@
 package com.green.member.application.admin;
 
-import com.green.member.application.MemberService;
+import com.green.member.application.member.MemberService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.web.bind.annotation.RequestMapping;
