+++ member-service/src/main/java/com/green/member/application/student/StudentRepository.java
@@ -1,6 +1,6 @@
 package com.green.member.application.student;
 
-import com.green.member.entity.Student;
+import com.green.member.entity.student.Student;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface StudentRepository extends JpaRepository<Student, Long> {
