+++ member-service/src/main/java/com/green/member/entity/cache/MajorCache.java
@@ -1,4 +1,4 @@
-package com.green.member.entity;
+package com.green.member.entity.cache;
 
 import jakarta.persistence.*;
 import lombok.*;
