+++ member-service/src/main/java/com/green/member/entity/student/MajorRequest.java
@@ -1,23 +1,24 @@
-package com.green.member.entity;
+package com.green.member.entity.student;
 
-import com.green.common.entity.CreatedAt;
 import com.green.common.entity.CreatedUpdatedAt;
 import com.green.common.enumcode.EnumApprovalStatus;
-import com.green.common.enumcode.EnumChangeType;
-import com.green.member.enumcode.EnumStatusRequestType;
+import com.green.member.enumcode.EnumMajorRequestType;
 import io.hypersistence.utils.hibernate.id.Tsid;
 import jakarta.persistence.*;
+import jakarta.validation.constraints.Digits;
 import lombok.*;
+
+import java.math.BigDecimal;
+
 @Entity
-@Table(name = "status_request")
+@Table(name = "major_request")
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
 @Builder
-public class StatusRequest extends CreatedUpdatedAt {
+public class MajorRequest extends CreatedUpdatedAt {
 
-    @Id
-    @Tsid
+    @Id @Tsid
     @Column(name = "request_id", nullable = false)
     private Long requestId;
 
@@ -26,13 +27,20 @@ public class StatusRequest extends CreatedUpdatedAt {
     private Student student;
 
     @Column(name = "type", nullable = false, length = 20)
-    private EnumStatusRequestType type;
+    private EnumMajorRequestType type;
 
-    @Column(name = "reason", nullable = false)
+    @Column(name = "target_major_id", nullable = false)
+    private Long targetMajorId;
+
+    @Column(name = "reason")
     private String reason;
 
     @Column(name = "file")
     private String file;
+
+    @Column(name = "gpa", nullable = false)
+    @Digits(integer = 1, fraction = 2)
+    private BigDecimal gpa;
 
     @Column(name = "status", nullable = false, length = 20)
     @Builder.Default
