+++ member-service/src/main/java/com/green/member/entity/student/StudentHistory.java
@@ -1,7 +1,7 @@
-package com.green.member.entity;
+package com.green.member.entity.student;
 
 import com.green.common.entity.CreatedAt;
-import com.green.member.enumcode.EnumAdminStatus;
+import com.green.common.enumcode.EnumStudentStatus;
 import io.hypersistence.utils.hibernate.id.Tsid;
 import jakarta.persistence.*;
 import lombok.*;
@@ -9,29 +9,29 @@ import lombok.*;
 import java.time.LocalDate;
 
 @Entity
-@Table(name = "admin_history")
+@Table(name = "student_history")
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
 @Builder
-public class AdminHistory extends CreatedAt {
+public class StudentHistory extends CreatedAt {
 
     @Id @Tsid
     @Column(name = "history_id", nullable = false)
     private Long historyId;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "admin_code", nullable = false)
-    private Admin admin;
+    @JoinColumn(name = "student_code", nullable = false)
+    private Student student;
 
     @Column(name = "change_type", nullable = false, length = 20)
-    private String changeType; //휴직, 복직
+    private String changeType; // 신입학, 휴학, 복학, 자퇴
 
     @Column(name = "old_status", nullable = false, length = 20)
-    private EnumAdminStatus oldStatus;
+    private EnumStudentStatus oldStatus;
 
     @Column(name = "new_status", nullable = false, length = 20)
-    private EnumAdminStatus newStatus;
+    private EnumStudentStatus newStatus;
 
     @Column(name = "start_date")
     private LocalDate startDate;
@@ -39,9 +39,17 @@ public class AdminHistory extends CreatedAt {
     @Column(name = "end_date")
     private LocalDate endDate;
 
+    // 휴학 시에만
+    @Column(name = "return_year")
+    private Integer returnYear;
+
+    // 휴학 시에만
+    @Column(name = "return_semester")
+    private Integer returnSemester;
+
     @Column(name = "reason")
     private String reason;
 
-    @Column(name = "updator_code")
+    @Column(name = "updator_code", nullable = false)
     private Long updatorCode;
 }
