+++ member-service/src/main/java/com/green/member/repository/MajorCacheRepository.java
@@ -1,6 +1,6 @@
 package com.green.member.repository;
 
-import com.green.member.entity.MajorCache;
+import com.green.member.entity.cache.MajorCache;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MajorCacheRepository extends JpaRepository<MajorCache, Long> {
